Add a last chance condition to Complex.parseTo
This is ridiculously complicated, and even so its ridiculously brittle. In the future, it would be wise to seek a better system, one that actually has some mathematical understanding of what its parsing. Oh well.

For the time being, however, this will intercept any "ni" form numbers that bypass other checks, and keep them from generating any exceptions. It should put at least one crash and a bug or two to rest.