Make complex data fields immutable
This is a combination of 2 commits. These commits are merged so that tools
like bisect work correctly.
The first commit's message is:

Start making Complex immutable

Complex needs to be immutable to protect its constants. That includes its internal data.

This is the 2nd commit message:

Define getters for Complex real and imaginary

Also, change all MainActivity references to the data structures to refer to the getters.