Squash commit for parseComplex
This is a combination of 5 commits, containing the implementation of the
parseCoefficient processor. These commits were squashed so tools like
rebase would function properly.

The first commit's message is:
Properly comment parseComplex

Also start working on using parseCoefficient.

This commit doesn't work for numbers that include pi or e.

This is the 2nd commit message:

Add the simplest case for parseCoefficient processing

This is the 3rd commit message:

Add the second-simplest case

Also add some basic error checking to parseCoefficient, necessary for this

This is the 4th commit message:

Add two more simple cases

Also slightly change the code for the last one.

This is the 5th commit message:

Implement the last case