Remove an unused method
This has been hanging around since
7bf22f7592986ff4a24333e09c2b0269e66fa3a0 (and, since that's the start of
version control for this application, since long before I made that
commit), just in case I found a use for it.

I haven't yet found a use for it.

Since it's going to stay in git for all time, I'm removing it. If I ever
have a chance to make use of it, I know where to find it.